LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-canvas-tainting-double-write.https-expected.txt


PASS canvas is tainted after writing both a non-opaque image and an opaque image from the same URL 
